Introduces an optional parameter, "required", for serial port config
If in the json config file required is set to true for a serial port,
the initialization of CommandManager will stop if that port is not
available.

This results in a more meaningful error rather than a warning likely to
generate an error later on in the initialization when the devices on
that arduino will not be found.

Old behaviour remains the default. This is useful when udev rules are
adopted in linux to identify connected Arduinos. When multiple ports are
blindly tested the old behaviour is a necessity (e.g. /dev/ttyACM[0-n])