Introduces initial_angle parameter for servos
This is important because otherwise all the servos are initialized to an
initial angle of 0 that may be unsuitable, depending on application.

Furthermore it adds min and max limits as parameters to __init__ so
that they can be set in the configuration JSON file, in the relevant
device "config" section. See example below:

"servo1": {
      "command_id": "servo1",
      "config": {
        "initial_angle": 90,
        "min_limit": 45,
        "max_limit": 135
      }
    }