Rename set_speed() into _set_speed() in commandlinearaccelstepper and commandaccelstepper
Ok so this might be confusing...

set_speed() sets the speed in the Arduino before actually moving the motor [e.g. in move(), move_to() or home()]
as such, it should not be called directly.

set_running_speed sets the value of the running_speed variable, used for set_speed before normal movements
set_homing_speed sets the value of the homing_speed variable, used for set_speed before homing movements in home()
set_max_speed sets the max allowed speed directly in the Arduino

set_speed should not be used because in every "moving" call (either move_to or move), set_speed will be called before
actually start the movement using self.running_speed as parameter (eventually with reversed sign if reverse_direction=True).
Similarly within the home() command, set_speed is called with self.homing_speed.
So, calling set_speed before a move() is pointless and set_running_speed should be called instead.
set_speed can still be meaningfully used to change the running speed of a motor while it is moving, but it is a rather
rare event.

This is so confusing that it was even wrongly showcased in the demo.py files.