Adds simulation parameter to CommandManager.from_configfile()
This is needed for simulation mode in DropletCollider as the commanduino
configuration is stored in a JSON file but simulation is set dynamically.

In this way if `simulation` is passed alongside the file path in
`from_configfile` the corresponding CommandManager parameter will be
set correctly.
Edit the file clearly is not a valid option.